fix ppp as matlab

+++ examples/at_focused_bowl_3D/at_focused_bowl_3D.py
@@ -73,7 +73,7 @@ grid_spacing_meters = Vector([dx, dx, dx])
 kgrid = kWaveGrid(grid_size_points, grid_spacing_meters)
 
 # compute points per temporal period
-ppp: float = np.round(ppw / cfl)
+ppp: int = round(ppw / cfl)
 
 # compute corresponding time spacing
 dt: float = 1.0 / (ppp * source_f0)
